requirements.txt
Flask>=2.0
requests>=2.0
dnspython>=2.0